+++ config/common.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Yiisoft\Definitions\Reference;
+use Yiisoft\Translator\MessageReaderInterface;
 use Yiisoft\Translator\TranslatorInterface;
 use Yiisoft\Translator\Translator;
 use Yiisoft\Translator\CategorySource;
+use Yiisoft\Translator\MessageFormatterInterface;
+use Yiisoft\Translator\SimpleMessageFormatter;
 
 /** @var array $params */
 
@@ -18,6 +21,16 @@ return [
     //         'name' => $params['yiisoft/translator']['defaultCategory'],
     //     ],
     // ],
+
+    'DefaultCategorySource' => static function (\Psr\Container\ContainerInterface $container) use ($params) {
+        return new CategorySource(
+            $params['yiisoft/translator']['defaultCategory'],
+            $container->get(MessageReaderInterface::class),
+            $container->has(MessageFormatterInterface::class)
+                ? $container->get(MessageFormatterInterface::class)
+                : new SimpleMessageFormatter(),
+        );
+    },
 
     TranslatorInterface::class => [
         'class' => Translator::class,