+++ config/params.php
@@ -10,6 +10,7 @@ return [
         'fallbackLocale' => null,
         'defaultCategory' => 'app',
         'categorySources' => [
+            Reference::to('DefaultCategorySource'),
             // You can add categories from your application and additional modules using `Reference::to` below
             // Reference::to(ApplicationCategorySource::class),
             // Reference::to(MyModuleCategorySource::class),